# Refactor Structural Block and Avoid Building Hash Table when Copying Map Block