--- a/presto-main/src/main/java/com/facebook/presto/operator/aggregation/DoubleHistogramAggregation.java
+++ b/presto-main/src/main/java/com/facebook/presto/operator/aggregation/DoubleHistogramAggregation.java
@@ -14,7 +14,6 @@
 package com.facebook.presto.operator.aggregation;
 
 import com.facebook.presto.operator.aggregation.state.DoubleHistogramStateSerializer;
-import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
 import com.facebook.presto.spi.function.AccumulatorState;
 import com.facebook.presto.spi.function.AccumulatorStateMetadata;
@@ -96,13 +95,13 @@ public final class DoubleHistogramAggregation
         }
         else {
             Map<Double, Double> value = state.get().getBuckets();
-            BlockBuilder blockBuilder = DoubleType.DOUBLE.createBlockBuilder(null, value.size() * 2);
+
+            BlockBuilder entryBuilder = out.beginBlockEntry();
             for (Map.Entry<Double, Double> entry : value.entrySet()) {
-                DoubleType.DOUBLE.writeDouble(blockBuilder, entry.getKey());
-                DoubleType.DOUBLE.writeDouble(blockBuilder, entry.getValue());
+                DoubleType.DOUBLE.writeDouble(entryBuilder, entry.getKey());
+                DoubleType.DOUBLE.writeDouble(entryBuilder, entry.getValue());
             }
-            Block block = blockBuilder.build();
-            out.appendStructure(block);
+            out.closeEntry();
         }
     }
 }