--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractMapBlock.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractMapBlock.java
@@ -224,16 +224,15 @@ public abstract class AbstractMapBlock
             }
             else {
                 getKeys().writePositionTo(i, entryBuilder);
-                entryBuilder.closeEntry();
             }
             if (getValues().isNull(i)) {
                 entryBuilder.appendNull();
             }
             else {
                 getValues().writePositionTo(i, entryBuilder);
-                entryBuilder.closeEntry();
             }
         }
+        blockBuilder.closeEntry();
     }
 
     @Override