--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/ArrayBlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/ArrayBlockBuilder.java
@@ -139,11 +139,12 @@ public class ArrayBlockBuilder
     }
 
     @Override
-    public BlockBuilder writeObject(Object value)
+    public BlockBuilder appendStructure(Object value)
     {
         if (currentEntryOpened) {
             throw new IllegalStateException("Expected current entry to be closed but was opened");
         }
+        currentEntryOpened = true;
 
         Block block = (Block) value;
         for (int i = 0; i < block.getPositionCount(); i++) {
@@ -156,7 +157,7 @@ public class ArrayBlockBuilder
             }
         }
 
-        currentEntryOpened = true;
+        closeEntry();
         return this;
     }