--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/ArrayBlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/ArrayBlockBuilder.java
@@ -152,7 +152,6 @@ public class ArrayBlockBuilder
             }
             else {
                 block.writePositionTo(i, values);
-                values.closeEntry();
             }
         }