--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/BlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/BlockBuilder.java
@@ -89,7 +89,7 @@ public interface BlockBuilder
     /**
      * Append a struct to the block and close the entry.
      */
-    default BlockBuilder appendStructure(Object value)
+    default BlockBuilder appendStructure(Block value)
     {
         throw new UnsupportedOperationException(getClass().getName());
     }