--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/ByteArrayBlock.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/ByteArrayBlock.java
@@ -119,6 +119,7 @@ public class ByteArrayBlock
     {
         checkReadablePosition(position);
         blockBuilder.writeByte(values[position + arrayOffset]);
+        blockBuilder.closeEntry();
     }
 
     @Override