--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/ByteArrayBlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/ByteArrayBlockBuilder.java
@@ -179,6 +179,7 @@ public class ByteArrayBlockBuilder
     {
         checkReadablePosition(position);
         blockBuilder.writeByte(values[position]);
+        blockBuilder.closeEntry();
     }
 
     @Override