--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/LongArrayBlock.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/LongArrayBlock.java
@@ -166,6 +166,7 @@ public class LongArrayBlock
     {
         checkReadablePosition(position);
         blockBuilder.writeLong(values[position + arrayOffset]);
+        blockBuilder.closeEntry();
     }
 
     @Override