--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/LongArrayBlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/LongArrayBlockBuilder.java
@@ -227,6 +227,7 @@ public class LongArrayBlockBuilder
     {
         checkReadablePosition(position);
         blockBuilder.writeLong(values[position]);
+        blockBuilder.closeEntry();
     }
 
     @Override