--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/MapBlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/MapBlockBuilder.java
@@ -262,7 +262,7 @@ public class MapBlockBuilder
     }
 
     @Override
-    public BlockBuilder writeObject(Object value)
+    public BlockBuilder appendStructure(Object value)
     {
         if (currentEntryOpened) {
             throw new IllegalStateException("Expected current entry to be closed but was opened");
@@ -290,6 +290,8 @@ public class MapBlockBuilder
                 valueBlockBuilder.closeEntry();
             }
         }
+
+        closeEntry();
         return this;
     }