--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/MapBlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/MapBlockBuilder.java
@@ -282,14 +282,12 @@ public class MapBlockBuilder
             }
             else {
                 block.writePositionTo(i, keyBlockBuilder);
-                keyBlockBuilder.closeEntry();
             }
             if (block.isNull(i + 1)) {
                 valueBlockBuilder.appendNull();
             }
             else {
                 block.writePositionTo(i + 1, valueBlockBuilder);
-                valueBlockBuilder.closeEntry();
             }
         }