--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/RowBlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/RowBlockBuilder.java
@@ -232,7 +232,6 @@ public class RowBlockBuilder
             }
             else {
                 block.writePositionTo(i, fieldBlockBuilders[i]);
-                fieldBlockBuilders[i].closeEntry();
             }
         }