--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/RowBlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/RowBlockBuilder.java
@@ -240,6 +240,30 @@ public class RowBlockBuilder
     }
 
     @Override
+    public BlockBuilder appendStructureInternal(Block block, int position)
+    {
+        if (!(block instanceof AbstractRowBlock)) {
+            throw new IllegalArgumentException();
+        }
+
+        AbstractRowBlock rowBlock = (AbstractRowBlock) block;
+        BlockBuilder entryBuilder = this.beginBlockEntry();
+
+        int fieldBlockOffset = rowBlock.getFieldBlockOffset(position);
+        for (int i = 0; i < rowBlock.numFields; i++) {
+            if (rowBlock.getFieldBlocks()[i].isNull(fieldBlockOffset)) {
+                entryBuilder.appendNull();
+            }
+            else {
+                rowBlock.getFieldBlocks()[i].writePositionTo(fieldBlockOffset, entryBuilder);
+            }
+        }
+
+        closeEntry();
+        return this;
+    }
+
+    @Override
     public BlockBuilder newBlockBuilderLike(BlockBuilderStatus blockBuilderStatus)
     {
         int newSize = calculateBlockResetSize(getPositionCount());