--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/ShortArrayBlock.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/ShortArrayBlock.java
@@ -119,6 +119,7 @@ public class ShortArrayBlock
     {
         checkReadablePosition(position);
         blockBuilder.writeShort(values[position + arrayOffset]);
+        blockBuilder.closeEntry();
     }
 
     @Override