--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/ShortArrayBlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/ShortArrayBlockBuilder.java
@@ -180,6 +180,7 @@ public class ShortArrayBlockBuilder
     {
         checkReadablePosition(position);
         blockBuilder.writeShort(values[position]);
+        blockBuilder.closeEntry();
     }
 
     @Override