--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/SingleArrayBlockWriter.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/SingleArrayBlockWriter.java
@@ -98,9 +98,9 @@ public class SingleArrayBlockWriter
     }
 
     @Override
-    public BlockBuilder appendStructure(Object value)
+    public BlockBuilder appendStructure(Block block)
     {
-        blockBuilder.appendStructure(value);
+        blockBuilder.appendStructure(block);
         entryAdded();
         return this;
     }