--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/SingleArrayBlockWriter.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/SingleArrayBlockWriter.java
@@ -106,6 +106,14 @@ public class SingleArrayBlockWriter
     }
 
     @Override
+    public BlockBuilder appendStructureInternal(Block block, int position)
+    {
+        blockBuilder.appendStructureInternal(block, position);
+        entryAdded();
+        return this;
+    }
+
+    @Override
     public BlockBuilder beginBlockEntry()
     {
         return blockBuilder.beginBlockEntry();