--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/SingleMapBlockWriter.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/SingleMapBlockWriter.java
@@ -154,6 +154,19 @@ public class SingleMapBlockWriter
     }
 
     @Override
+    public BlockBuilder appendStructureInternal(Block block, int position)
+    {
+        if (writeToValueNext) {
+            valueBlockBuilder.appendStructureInternal(block, position);
+        }
+        else {
+            keyBlockBuilder.appendStructureInternal(block, position);
+        }
+        entryAdded();
+        return this;
+    }
+
+    @Override
     public BlockBuilder beginBlockEntry()
     {
         BlockBuilder result;