--- a/presto-spi/src/main/java/com/facebook/presto/spi/type/ArrayType.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/type/ArrayType.java
@@ -187,7 +187,7 @@ public class ArrayType
     @Override
     public void writeObject(BlockBuilder blockBuilder, Object value)
     {
-        blockBuilder.writeObject(value).closeEntry();
+        blockBuilder.appendStructure(value);
     }
 
     @Override