Benchmark:

Before
```
Benchmark                          (mapSize)  Mode  Cnt     Score    Error  Units
BenchmarkMapCopy.benchmarkMapCopy          1  avgt  100   141.925 ±  0.830  ns/op
BenchmarkMapCopy.benchmarkMapCopy          2  avgt  100   277.466 ±  2.169  ns/op
BenchmarkMapCopy.benchmarkMapCopy          4  avgt  100   541.838 ±  5.143  ns/op
BenchmarkMapCopy.benchmarkMapCopy          8  avgt  100  1127.118 ± 39.768  ns/op
BenchmarkMapCopy.benchmarkMapCopy         16  avgt  100  2212.371 ± 84.363  ns/op
```

After
```
Benchmark                          (mapSize)  Mode  Cnt     Score    Error  Units
BenchmarkMapCopy.benchmarkMapCopy          1  avgt  100    89.487 ±  0.736  ns/op
BenchmarkMapCopy.benchmarkMapCopy          2  avgt  100   157.988 ±  1.185  ns/op
BenchmarkMapCopy.benchmarkMapCopy          4  avgt  100   291.660 ±  1.935  ns/op
BenchmarkMapCopy.benchmarkMapCopy          8  avgt  100   589.937 ± 23.140  ns/op
BenchmarkMapCopy.benchmarkMapCopy         16  avgt  100  1178.167 ± 45.828  ns/op
```

It's 37% to 48% improvement.